measured execution time - did some code cleanup

+++ 1_helloworld/src/kernel_files/HelloWorld.cl
@@ -1,6 +1,6 @@
-__kernel void hello_kernel(__global const float *a, 
-						   __global const float *b, 
-						   __global       float *result)
+__kernel void hello_kernel(__global const double *a, 
+						   __global const double *b, 
+						   __global       double *result)
 {
     int gid = get_global_id(0);
     result[gid] = a[gid] + b[gid];
